requirements.txt
fastapi==0.114.0
mangum==0.17.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
